+++ src/tag.rs
@@ -1,69 +1,74 @@
 use crate::byte_order::ByteOrder;
+use crate::errors::Error;
+use crate::parser;
 
-pub type Name<'a> = Option<&'a str>;
+pub type Name = Option<String>;
 
 /// A tag is an individual part of the data tree. A tag consists of a name and
 /// a payload. The name is absent when it is used within a [List].
 ///
 /// The tag `TAG_End` is not available because it is handled by the program.
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub enum Tag<'a> {
+#[derive(Clone, Debug, PartialEq)]
+pub enum Tag {
     /// A signed integral type. Sometimes used for booleans.
-    Byte(Name<'a>, i8),
+    Byte(Name, i8),
 
     /// A signed integral type.
-    Short(Name<'a>, i16),
+    Short(Name, i16),
 
     /// A signed intefral type.
-    Int(Name<'a>, i32),
+    Int(Name, i32),
 
     /// A signed integral type.
-    Long(Name<'a>, i64),
+    Long(Name, i64),
 
     /// A signed floating point type.
-    Float(Name<'a>, f32),
+    Float(Name, f32),
 
     /// A signed floating point type.
-    Double(Name<'a>, f64),
+    Double(Name, f64),
 
     /// An array of bytes.
-    ByteArray(Name<'a>, &'a [i8]),
+    ByteArray(Name, Vec<i8>),
 
     /// A UTF-8 string.
-    String(Name<'a>, &'a str),
+    String(Name, String),
 
     /// A list of tag payloads, without tag IDs or names.
-    List(Name<'a>, &'a [Tag<'a>]),
+    List(Name, Vec<Tag>),
 
     /// A list of fully formed tags, including their IDs, names, and payloads.
-    Compound(Name<'a>, &'a [Tag<'a>]),
+    Compound(Name, Vec<Tag>),
 
     /// An array of [Tag::Int]s.
-    IntArray(Name<'a>, &'a [i32]),
+    IntArray(Name, Vec<i32>),
 
     /// An array of [Tag::Long]s.
-    LongArray(Name<'a>, &'a [i64]),
+    LongArray(Name, Vec<i64>),
 }
 
-impl Tag<'_> {
+impl Tag {
     /// Returns a tag represented as bytes.
-    pub fn as_bytes(&self, byte_order: ByteOrder) -> Vec<u8> {
+    pub fn to_bytes(&self, byte_order: ByteOrder) -> Result<Vec<u8>, Error> {
         let mut buf = vec![];
         buf.extend(self.bytes_id(byte_order));
-        buf.extend(
-            self.bytes_name(byte_order)
-                .expect("invalid structure: expected named tag"),
-        );
-        buf.extend(self.bytes_payload(byte_order));
-        buf
+        buf.extend(self.bytes_name(byte_order)?);
+        buf.extend(self.bytes_payload(byte_order)?);
+        Ok(buf)
+    }
+
+    /// Returns a tag from bytes.
+    #[cfg(feature = "read")]
+    pub fn from_bytes(bytes: &[u8], byte_order: ByteOrder) -> Result<Self, Error> {
+        Ok(parser::nbt(bytes, byte_order).expect("TODO").1)
     }
 
     /// Returns a vector of bytes with the length of the name and the name
     /// MUTF-8 encoded. This is an [std::result::Result::Err] when the name
     /// is absent.
-    fn bytes_name(&self, byte_order: ByteOrder) -> Result<Vec<u8>, ()> {
+    fn bytes_name(&self, byte_order: ByteOrder) -> Result<Vec<u8>, Error> {
         let mut buf = vec![];
-        match *self {
+        match self {
             Tag::Byte(name, _)
             | Tag::Short(name, _)
             | Tag::Int(name, _)
@@ -76,9 +81,9 @@ impl Tag<'_> {
             | Tag::Compound(name, _)
             | Tag::IntArray(name, _)
             | Tag::LongArray(name, _) => {
-                let n = name.ok_or(())?;
+                let n = name.as_ref().ok_or(Error::MissingName)?;
                 let n = mutf8::encode(n).into_owned();
-                let len: u16 = n.len().try_into().expect("name too big");
+                let len: u16 = n.len().try_into().map_err(|_| Error::StringTooBig)?;
                 buf.extend(byte_order.bytes(len));
                 buf.extend(n);
             }
@@ -107,19 +112,22 @@ impl Tag<'_> {
     }
 
     /// Returns the payload of the tag as a vector of bytes.
-    fn bytes_payload(&self, byte_order: ByteOrder) -> Vec<u8> {
+    fn bytes_payload(&self, byte_order: ByteOrder) -> Result<Vec<u8>, Error> {
         let mut buf = vec![];
-        match *self {
-            Tag::Byte(_, payload) => buf.extend(byte_order.bytes(payload)),
-            Tag::Short(_, payload) => buf.extend(byte_order.bytes(payload)),
-            Tag::Int(_, payload) => buf.extend(byte_order.bytes(payload)),
-            Tag::Long(_, payload) => buf.extend(byte_order.bytes(payload)),
-            Tag::Float(_, payload) => buf.extend(byte_order.bytes(payload)),
-            Tag::Double(_, payload) => buf.extend(byte_order.bytes(payload)),
+        match self {
+            Tag::Byte(_, payload) => buf.extend(byte_order.bytes(*payload)),
+            Tag::Short(_, payload) => buf.extend(byte_order.bytes(*payload)),
+            Tag::Int(_, payload) => buf.extend(byte_order.bytes(*payload)),
+            Tag::Long(_, payload) => buf.extend(byte_order.bytes(*payload)),
+            Tag::Float(_, payload) => buf.extend(byte_order.bytes(*payload)),
+            Tag::Double(_, payload) => buf.extend(byte_order.bytes(*payload)),
 
             Tag::ByteArray(_, payload) => {
                 // length of array
-                let len: i32 = payload.len().try_into().expect("byte array too big");
+                let len: i32 = payload
+                    .len()
+                    .try_into()
+                    .map_err(|_| Error::ByteArrayTooBig)?;
                 buf.extend(byte_order.bytes(len));
 
                 // content of array
@@ -129,8 +137,8 @@ impl Tag<'_> {
             }
 
             Tag::String(_, payload) => {
-                let string = mutf8::encode(payload);
-                let len: u16 = string.len().try_into().expect("string too big");
+                let string = mutf8::encode(payload.as_str());
+                let len: u16 = string.len().try_into().map_err(|_| Error::StringTooBig)?;
                 buf.extend(byte_order.bytes(len));
                 buf.extend(string.into_owned());
             }
@@ -142,25 +150,28 @@ impl Tag<'_> {
                 }
 
                 // length of list
-                let len: i32 = payload.len().try_into().expect("list too big");
+                let len: i32 = payload.len().try_into().map_err(|_| Error::ListTooBig)?;
                 buf.extend(byte_order.bytes(len));
 
                 // content of list
                 for byte in payload {
-                    buf.extend(byte.bytes_payload(byte_order));
+                    buf.extend(byte.bytes_payload(byte_order)?);
                 }
             }
 
             Tag::Compound(_, payload) => {
                 for tag in payload {
-                    buf.extend(tag.as_bytes(byte_order));
+                    buf.extend(tag.to_bytes(byte_order)?);
                 }
                 buf.extend(byte_order.bytes(0_i8));
             }
 
             Tag::IntArray(_, payload) => {
                 // length of array
-                let len: i32 = payload.len().try_into().expect("int array too big");
+                let len: i32 = payload
+                    .len()
+                    .try_into()
+                    .map_err(|_| Error::IntArrayTooBig)?;
                 buf.extend(byte_order.bytes(len));
 
                 // content of array
@@ -171,7 +182,10 @@ impl Tag<'_> {
 
             Tag::LongArray(_, payload) => {
                 // length of array
-                let len: i32 = payload.len().try_into().expect("long array too big");
+                let len: i32 = payload
+                    .len()
+                    .try_into()
+                    .map_err(|_| Error::LongArrayTooBig)?;
                 buf.extend(byte_order.bytes(len));
 
                 // content of array
@@ -180,7 +194,7 @@ impl Tag<'_> {
                 }
             }
         };
-        buf
+        Ok(buf)
     }
 }
 
@@ -192,16 +206,22 @@ impl Tag<'_> {
 /// use mcnbt::{Tag, nbt};
 ///
 /// assert_eq!(
-///     nbt!(Tag::Int(Some("foo"), 42), Tag::Long(Some("bar"), 12)),
+///     nbt![
+///         Tag::Int(Some("foo".to_string()), 42),
+///         Tag::Long(Some("bar".to_string()), 12)
+///     ],
 ///     Tag::Compound(
-///         Some(""),
-///         &[Tag::Int(Some("foo"), 42), Tag::Long(Some("bar"), 12)]
+///         Some("".to_string()),
+///         vec![
+///             Tag::Int(Some("foo".to_string()), 42),
+///             Tag::Long(Some("bar".to_string()), 12)
+///         ]
 ///     )
 /// );
 /// ```
 #[macro_export]
 macro_rules! nbt {
     ($($data:expr),* $(,)?) => {
-        Tag::Compound(Some(""), &[$($data),*])
+        Tag::Compound(Some(String::new()), vec![$($data),*])
     };
 }