+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Changelog
 
-## [0.2.0]
+## [unreleased]
+
+### Added
+
+- Add `mcnbt::errors` module.
+
+### Changed
+
+- **BREAKING**: `mcnbt::tag::Tag::as_bytes` has been renamed to `to_bytes` and now
+  returns a `Result`.
+- **BREAKING**: Changed types of `mcnbt::tag::Tag` variants.
+
+### Removed
+
+- **BREAKING**: Remove `mcnbt::FORMAT_VERSION`.
+
+
+## [0.2.0] - 2023-10-21
 
 ### Added
 
@@ -14,5 +31,6 @@
 _🍰 Initial release_
 
 
+[unreleased]: https://github.com/phoenixr-codes/mcnbt/compare/v0.2.0...HEAD
 [0.2.0]: https://github.com/phoenixr-codes/mcnbt/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/phoenixr-codes/mcnbt/releases/v0.1.0